Use more generic UnauthorizedProblem
In many cases, OAuthProblem or OAuthResponseProblem was being raised for
generic auth issues instead of for things related to OAuth. So, add a
new UnauthorizedProblem to handle all of those generic use cases
instead.

Also leave a note on the methods where it still makes sense to use the
OAuthResponseProblem.